test(ios): keep wait helpers under file budget
